renaming ./pkg/ to ./internal/
