+++ .github/workflows/build.yml
@@ -28,4 +28,4 @@ jobs:
       run: go build -tags dev -v ./cmd/server/main.go ./cmd/server/wire_gen.go
 
     - name: Test
-      run: go test -tags dev -v ./pkg/...
+      run: go test -tags dev -v ./internal/...
