+++ cmd/server/wire.go
@@ -3,9 +3,9 @@ package main
 import (
 	"github.com/google/wire"
 
-	"github.com/eduardogr/webser-go/pkg/api"
-	"github.com/eduardogr/webser-go/pkg/repository"
-	"github.com/eduardogr/webser-go/pkg/server"
+	"github.com/eduardogr/webser-go/internal/api"
+	"github.com/eduardogr/webser-go/internal/repository"
+	"github.com/eduardogr/webser-go/internal/server"
 )
 
 func InitializeServer() (*server.Server, error) {
